[FIXED] anim2 load

+++ source/NodeAnim2Spr.cpp
@@ -9,7 +9,7 @@ namespace simp
 NodeAnim2Spr::NodeAnim2Spr(ImportStream& is)
 {
 	sym = is.UInt32();
-	static_time = is.UInt16();
+	static_time = is.UInt32();
 }
 
 int NodeAnim2Spr::Size()
